gpui, zed: Add FrameRecorder and visual test harness for UI recording
Adds infrastructure for programmatically driving real Zed UI and exporting
the result as an animated GIF or PNG frame sequence.

- `gpui::FrameRecorder` accumulates rendered frames and exports via
  `export_gif` or `export_frames_to_directory`
- `VisualTestAppContext::record_frame` implicitly parks before capture
  so callers never need manual `run_until_parked` before recording
- `zed::visual_harness::setup_zed_for_visual_test` extracts the full
  subsystem init from `visual_test_runner` into a reusable harness,
  returning a `(WindowHandle<Workspace>, Arc<AppState>)`

No new Cargo dependencies — GIF encoding uses the `image` crate already
in the transitive dependency graph.

Release Notes:

- N/A
