+++ crates/gpui/src/gpui.rs
@@ -47,6 +47,8 @@ mod svg_renderer;
 mod tab_stop;
 mod taffy;
 #[cfg(any(test, feature = "test-support"))]
+pub mod recording;
+#[cfg(any(test, feature = "test-support"))]
 pub mod test;
 mod text_system;
 mod util;
@@ -115,6 +117,8 @@ pub use svg_renderer::*;
 pub(crate) use tab_stop::*;
 use taffy::TaffyLayoutEngine;
 pub use taffy::{AvailableSpace, LayoutId};
+#[cfg(any(test, feature = "test-support"))]
+pub use recording::{CapturedFrame, FrameRecorder};
 #[cfg(any(test, feature = "test-support"))]
 pub use test::*;
 pub use text_system::*;
