+++ crates/zed/src/visual_test_runner.rs
@@ -147,148 +147,22 @@ use constants::*;
 
 #[cfg(target_os = "macos")]
 fn run_visual_tests(project_path: PathBuf, update_baseline: bool) -> Result<()> {
-    // Create the visual test context with deterministic task scheduling
-    // Use real Assets so that SVG icons render properly
     let mut cx = VisualTestAppContext::with_asset_source(
         gpui_platform::current_platform(false),
         Arc::new(Assets),
     );
 
-    // Load embedded fonts (IBM Plex Sans, Lilex, etc.) so UI renders with correct fonts
+    let (workspace_window, app_state) =
+        crate::zed::visual_harness::setup_zed_for_visual_test(&mut cx)?;
+
+    // Bind keymaps specific to the visual test runner
     cx.update(|cx| {
-        Assets.load_fonts(cx).unwrap();
-    });
-
-    // Initialize settings store with real default settings (not test settings)
-    // Test settings use Courier font, but we want the real Zed fonts for visual tests
-    cx.update(|cx| {
-        settings::init(cx);
-    });
-
-    // Create AppState using the test initialization
-    let app_state = cx.update(|cx| init_app_state(cx));
-
-    // Set the global app state so settings_ui and other subsystems can find it
-    cx.update(|cx| {
-        AppState::set_global(app_state.clone(), cx);
-    });
-
-    // Initialize all Zed subsystems
-    cx.update(|cx| {
-        gpui_tokio::init(cx);
-        theme_settings::init(theme::LoadThemes::JustBase, cx);
-        client::init(&app_state.client, cx);
-        audio::init(cx);
-        workspace::init(app_state.clone(), cx);
-        release_channel::init(semver::Version::new(0, 0, 0), cx);
-        command_palette::init(cx);
-        editor::init(cx);
-        call::init(app_state.client.clone(), app_state.user_store.clone(), cx);
-        title_bar::init(cx);
-        project_panel::init(cx);
-        outline_panel::init(cx);
-        terminal_view::init(cx);
-        image_viewer::init(cx);
-        search::init(cx);
-        cx.set_global(workspace::PaneSearchBarCallbacks {
-            setup_search_bar: |languages, toolbar, window, cx| {
-                let search_bar = cx.new(|cx| search::BufferSearchBar::new(languages, window, cx));
-                toolbar.update(cx, |toolbar, cx| {
-                    toolbar.add_item(search_bar, window, cx);
-                });
-            },
-            wrap_div_with_search_actions: search::buffer_search::register_pane_search_actions,
-        });
-        prompt_store::init(cx);
-        let prompt_builder = prompt_store::PromptBuilder::load(app_state.fs.clone(), false, cx);
-        language_model::init(cx);
-        client::RefreshLlmTokenListener::register(
-            app_state.client.clone(),
-            app_state.user_store.clone(),
-            cx,
-        );
-        language_models::init(app_state.user_store.clone(), app_state.client.clone(), cx);
-        git_ui::init(cx);
-        project::AgentRegistryStore::init_global(
-            cx,
-            app_state.fs.clone(),
-            app_state.client.http_client(),
-        );
-        agent_ui::init(
-            app_state.fs.clone(),
-            prompt_builder,
-            app_state.languages.clone(),
-            true,
-            false,
-            cx,
-        );
-        settings_ui::init(cx);
-
-        // Load default keymaps so tooltips can show keybindings like "f9" for ToggleBreakpoint
-        // We load a minimal set of editor keybindings needed for visual tests
         cx.bind_keys([KeyBinding::new(
             "f9",
             editor::actions::ToggleBreakpoint,
             Some("Editor"),
         )]);
-
-        // Disable agent notifications during visual tests to avoid popup windows
-        agent_settings::AgentSettings::override_global(
-            agent_settings::AgentSettings {
-                notify_when_agent_waiting: NotifyWhenAgentWaiting::Never,
-                play_sound_when_agent_done: PlaySoundWhenAgentDone::Never,
-                ..agent_settings::AgentSettings::get_global(cx).clone()
-            },
-            cx,
-        );
     });
-
-    // Run until all initialization tasks complete
-    cx.run_until_parked();
-
-    // Open workspace window
-    let window_size = size(px(1280.0), px(800.0));
-    let bounds = Bounds {
-        origin: point(px(0.0), px(0.0)),
-        size: window_size,
-    };
-
-    // Create a project for the workspace
-    let project = cx.update(|cx| {
-        project::Project::local(
-            app_state.client.clone(),
-            app_state.node_runtime.clone(),
-            app_state.user_store.clone(),
-            app_state.languages.clone(),
-            app_state.fs.clone(),
-            None,
-            project::LocalProjectFlags {
-                init_worktree_trust: false,
-                ..Default::default()
-            },
-            cx,
-        )
-    });
-
-    let workspace_window: WindowHandle<Workspace> = cx
-        .update(|cx| {
-            cx.open_window(
-                WindowOptions {
-                    window_bounds: Some(WindowBounds::Windowed(bounds)),
-                    focus: false,
-                    show: false,
-                    ..Default::default()
-                },
-                |window, cx| {
-                    cx.new(|cx| {
-                        Workspace::new(None, project.clone(), app_state.clone(), window, cx)
-                    })
-                },
-            )
-        })
-        .context("Failed to open workspace window")?;
-
-    cx.run_until_parked();
 
     // Add the test project as a worktree
     let add_worktree_task = workspace_window
@@ -961,48 +835,6 @@ cargo run
     std::fs::write(project_path.join("README.md"), readme).expect("Failed to write README.md");
 }
 
-#[cfg(target_os = "macos")]
-fn init_app_state(cx: &mut App) -> Arc<AppState> {
-    use fs::Fs;
-    use node_runtime::NodeRuntime;
-    use session::Session;
-    use settings::SettingsStore;
-
-    if !cx.has_global::<SettingsStore>() {
-        let settings_store = SettingsStore::test(cx);
-        cx.set_global(settings_store);
-    }
-
-    // Use the real filesystem instead of FakeFs so we can access actual files on disk
-    let fs: Arc<dyn Fs> = Arc::new(fs::RealFs::new(None, cx.background_executor().clone()));
-    <dyn Fs>::set_global(fs.clone(), cx);
-
-    let languages = Arc::new(language::LanguageRegistry::test(
-        cx.background_executor().clone(),
-    ));
-    let clock = Arc::new(clock::FakeSystemClock::new());
-    let http_client = http_client::FakeHttpClient::with_404_response();
-    let client = client::Client::new(clock, http_client, cx);
-    let session = cx.new(|cx| session::AppSession::new(Session::test(), cx));
-    let user_store = cx.new(|cx| client::UserStore::new(client.clone(), cx));
-    let workspace_store = cx.new(|cx| workspace::WorkspaceStore::new(client.clone(), cx));
-
-    theme_settings::init(theme::LoadThemes::JustBase, cx);
-    client::init(&client, cx);
-
-    let app_state = Arc::new(AppState {
-        client,
-        fs,
-        languages,
-        user_store,
-        workspace_store,
-        node_runtime: NodeRuntime::unavailable(),
-        build_window_options: |_, _| Default::default(),
-        session,
-    });
-    AppState::set_global(app_state.clone(), cx);
-    app_state
-}
 
 /// Runs visual tests for breakpoint hover states in the editor gutter.
 ///
